requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
torch>=2.6.0
transformers==4.37.0
Pillow==9.5.0
pinecone-client==3.0.0
python-multipart==0.0.6
jinja2==3.1.3 